backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.0
httpx==0.27.0
openai==1.35.0
pydantic==2.7.1
pydantic-settings==2.2.0
sqlalchemy==2.0.23
python-multipart==0.0.9
python-dateutil==2.8.2
psycopg2-binary==2.9.9  # For Neon PostgreSQL